1.修改工程信息收藏功能异常bug 2.修改基建物资预定选择存放地数量输入过多闪退bug 3.修改物资取消预定页面不刷新异常

+++ app/src/main/java/com/shuangduan/zcy/model/api/ApiService.java
@@ -99,6 +99,7 @@ import com.shuangduan.zcy.model.bean.WXLoginVerificationBean;
 import com.shuangduan.zcy.model.bean.WithdrawBean;
 import com.shuangduan.zcy.model.bean.WithdrawRecordBean;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import io.reactivex.Flowable;
@@ -1247,7 +1248,7 @@ public interface ApiService {
     Flowable<BaseResponse<MaterialAddBean>> getAddMaterial(
             @Field("user_id") int userId,
             @Field("material_id") int material_id,
-            @Field("num") int num
+            @Field("num") BigInteger num
     );
 
     //基建物资提交预订单 删除存放地
